port xmlRepo and releaseRepo to use postgres
