+++ src/db.ts
@@ -1,19 +1,14 @@
-import sql, { Database, Query } from '@radically-straightforward/sqlite'
+import sql, { Database } from '@radically-straightforward/sqlite'
 import { Statement } from 'better-sqlite3'
 import { randomBytes } from 'node:crypto'
-import postgres from 'postgres'
-import { DDEXRelease, DDEXReleaseIds } from './parseDelivery'
+import { DDEXRelease } from './parseDelivery'
 import { dataDir } from './sources'
 import { lowerAscii } from './util'
 
+export { assetRepo, releaseRepo, xmlRepo } from './pg'
+
 const dbLocation = process.env.SQLITE_URL || `${dataDir}/ddex.db`
 const db = new Database(dbLocation)
-
-const pgsql = postgres({
-  port: 40111,
-  user: 'postgres',
-  pass: 'example',
-})
 
 db.pragma('journal_mode = WAL')
 db.pragma('synchronous = NORMAL')
@@ -115,45 +110,6 @@ create table if not exists s3markers (
   sql`delete from users where id = 'lzAWJyO'`,
   sql`alter table xmls drop column xmlText`
 )
-
-export async function pgMigrate() {
-  await pgsql`
-  CREATE TABLE IF NOT EXISTS xmls (
-    "source" text not null,
-    "xmlUrl" text primary key,
-    "messageTimestamp" text not null,
-    "createdAt" timestamptz DEFAULT CURRENT_TIMESTAMP
-  );`
-
-  await pgsql`
-  CREATE TABLE IF NOT EXISTS releases (
-    "source" text not null,
-    "key" text primary key,
-    "ref" text,
-    "xmlUrl" text,
-    "messageTimestamp" text,
-    "json" jsonb,
-    "status" text not null,
-
-    "entityType" text,
-    "entityId" text,
-    "blockHash" text,
-    "blockNumber" integer,
-    "publishedAt" timestamptz,
-
-    "publishErrorCount" integer default 0,
-    "lastPublishError" text,
-
-    "createdAt" timestamptz DEFAULT CURRENT_TIMESTAMP,
-    "updatedAt" timestamptz,
-    "releaseType" text,
-    "releaseDate" text,
-    "numCleared" int,
-    "numNotCleared" int,
-    "prependArtist" boolean
-  );
-  `
-}
 
 export type XmlRow = {
   source: string
@@ -266,296 +222,6 @@ export const userRepo = {
         return u.id
       }
     }
-  },
-}
-
-//
-// xml repo
-//
-
-export const xmlRepo = {
-  all(cursor: string) {
-    return db.all<XmlRow>(sql`
-      select * from xmls
-      where xmlUrl > ${cursor}
-      order by xmlUrl
-      limit 1000
-      `)
-  },
-
-  get(xmlUrl: string) {
-    return db.get<XmlRow>(sql`select * from xmls where xmlUrl = ${xmlUrl}`)
-  },
-
-  find(query: string) {
-    const xmls = db.all<XmlRow>(
-      sql`
-        select * from xmls
-        where xmlUrl like '%' || ${query} || '%'
-        order by messageTimestamp desc`
-    )
-    return xmls
-  },
-
-  async upsert(row: Partial<XmlRow>) {
-    dbUpsert('xmls', row)
-    await pgUpsert('xmls', 'xmlUrl', row)
-  },
-}
-
-//
-// Resource repo
-//
-
-export type AssetRow = {
-  releaseId: string
-  ref: string
-  xmlUrl: string
-  filePath: string
-  fileName: string
-}
-
-export const assetRepo = {
-  get(source: string, releaseId: string, ref: string) {
-    return db.get<AssetRow>(
-      sql`select * from assets where source = ${source} and releaseId = ${releaseId} and ref = ${ref}`
-    )
-  },
-}
-
-//
-// release repo
-//
-
-type FindReleaseParams = {
-  pendingPublish?: boolean
-  status?: string
-  source?: string
-  limit?: number
-  offset?: number
-  search?: string
-  cleared?: boolean
-}
-
-type StatsRow = {
-  source: string
-  count: number
-}
-
-export const releaseRepo = {
-  // todo: this is incomplete, and I'm not sure how to order which ID to use first.
-  //   go version used xml file name
-  //   but a single file can contain multiple releases
-  //   so still need a way to pick an identifier, right?
-  chooseReleaseId(releaseIds: DDEXReleaseIds) {
-    const key = releaseIds.isrc || releaseIds.icpn || releaseIds.grid
-    if (!key) {
-      const msg = `failed to chooseReleaseId: ${JSON.stringify(releaseIds)}`
-      console.log(msg)
-      throw new Error(msg)
-    }
-    return key
-  },
-
-  stats() {
-    return db.all<StatsRow>(
-      sql`select source, count(*) count from releases group by 1`
-    )
-  },
-
-  all(params?: FindReleaseParams) {
-    params ||= {}
-    const rows = db.all<ReleaseRow>(sql`
-      select * from releases
-      where 1=1
-
-      -- pending publish
-      $${ifdef(
-        params.pendingPublish,
-        sql`
-          and status in (
-            ${ReleaseProcessingStatus.PublishPending},
-            ${ReleaseProcessingStatus.Failed},
-            ${ReleaseProcessingStatus.DeletePending}
-          )
-          and publishErrorCount < 5 `
-      )}
-
-      $${ifdef(params.status, sql` and status = ${params.status} `)}
-
-      $${ifdef(params.source, sql` and source = ${params.source} `)}
-
-      $${ifdef(params.search, sql` and json like '%' || ${params.search} || '%' `)}
-
-      $${ifdef(params.cleared, sql` and numCleared > 0 `)}
-
-      order by messageTimestamp desc
-
-      $${ifdef(params.limit, sql` limit ${params.limit} `)}
-      $${ifdef(params.offset, sql` offset ${params.offset} `)}
-    `)
-
-    for (const row of rows) {
-      if (row.json) row._parsed = JSON.parse(row.json)
-    }
-    return rows
-  },
-
-  rawSelect(q: Query) {
-    const rows = db.all<ReleaseRow>(q)
-    for (const row of rows) {
-      if (row.json) row._parsed = JSON.parse(row.json)
-    }
-    return rows
-  },
-
-  get(key: string) {
-    const row = db.get<ReleaseRow>(
-      sql`select * from releases where key = ${key}`
-    )
-    if (!row) return
-    if (row.json) row._parsed = JSON.parse(row.json)
-    return row
-  },
-
-  update(r: Partial<ReleaseRow>) {
-    dbUpdate('releases', 'key', r)
-  },
-
-  upsert: async (
-    source: string,
-    xmlUrl: string,
-    messageTimestamp: string,
-    release: DDEXRelease
-  ) => {
-    const key = releaseRepo.chooseReleaseId(release.releaseIds)
-    const prior = releaseRepo.get(key)
-
-    // skip TrackRelease, since we only want main releases
-    if (release.releaseType == 'TrackRelease') {
-      return
-    }
-
-    // if prior exists and is newer, skip
-    if (prior && prior.messageTimestamp > messageTimestamp) {
-      console.log(`skipping ${xmlUrl} because ${key} is newer`)
-      return
-    }
-
-    if (!release.audiusUser) {
-      release.problems.push(`NoUser`)
-    } else {
-      const idx = release.problems.indexOf(`NoUser`)
-      if (idx != -1) {
-        release.problems.splice(idx, 1)
-      }
-    }
-
-    const json = JSON.stringify(release)
-
-    // if same xmlUrl + json, skip
-    // may want some smarter json compare here
-    // if this is causing spurious sdk updates to be issued
-    // if (prior && prior.xmlUrl == xmlUrl && prior.json == json) {
-    //   return
-    // }
-
-    let status: ReleaseRow['status'] = release.problems.length
-      ? ReleaseProcessingStatus.Blocked
-      : ReleaseProcessingStatus.PublishPending
-
-    // if prior is published and latest version has no deal,
-    // treat as takedown
-    if (prior?.entityId && release.deals.length == 0) {
-      status = ReleaseProcessingStatus.DeletePending
-    }
-
-    // pull out original resource URLs
-    // so if an update comes in we can still resolve the original file
-    for (const r of [...release.soundRecordings, ...release.images]) {
-      if (r.ref && r.filePath && r.fileName) {
-        dbUpsert('assets', {
-          source: source,
-          releaseId: key,
-          ref: r.ref,
-          xmlUrl: xmlUrl,
-          filePath: r.filePath,
-          fileName: r.fileName,
-        })
-
-        // await pgUpsert('')
-      }
-    }
-
-    const data = {
-      source,
-      key,
-      status,
-      ref: release.ref,
-      releaseType: release.releaseType,
-      releaseDate: release.releaseDate,
-      xmlUrl,
-      messageTimestamp,
-      json,
-      updatedAt: new Date().toISOString(),
-    } as Partial<ReleaseRow>
-
-    dbUpsert('releases', data)
-
-    await pgUpsert('releases', 'key', data)
-  },
-
-  markPrependArtist(key: string, prependArtist: boolean) {
-    db.run(sql`
-      update releases set
-      prependArtist=${prependArtist ? 'TRUE' : ''}
-      where key = ${key}
-    `)
-  },
-
-  markForDelete: db.transaction(
-    (
-      source: string,
-      xmlUrl: string,
-      messageTimestamp: string,
-      releaseIds: DDEXReleaseIds
-    ) => {
-      // here we do PK lookup using the "best" id
-      // but we may need to try to find by all the different releaseIds
-      // if it's not consistent
-      const key = releaseRepo.chooseReleaseId(releaseIds)
-      const prior = releaseRepo.get(key)
-
-      if (!prior) {
-        console.log(`got purge release but no prior ${key}`)
-        return
-      }
-
-      if (prior.messageTimestamp >= messageTimestamp) {
-        console.log(`skipping delete ${key}`)
-        return
-      }
-
-      releaseRepo.update({
-        key,
-        status: ReleaseProcessingStatus.DeletePending,
-        source,
-        xmlUrl,
-        messageTimestamp,
-      })
-    }
-  ),
-
-  addPublishError(key: string, err: Error) {
-    const status = ReleaseProcessingStatus.Failed
-    const errText = err.stack || err.toString()
-    db.run(sql`
-      update releases set
-        status=${status},
-        lastPublishError=${errText},
-        publishErrorCount=publishErrorCount+1
-      where key = ${key}
-    `)
   },
 }
 
@@ -687,25 +353,6 @@ function dbUpsert(table: string, data: Record<string, any>) {
     insert into ${table} (${fields}) values (${qs})
     on conflict do update set ${excludes}`
   return toStmt(rawSql).run(...Object.values(data))
-}
-
-async function pgUpsert(
-  table: string,
-  pkField: string,
-  data: Record<string, any>
-) {
-  await pgsql.begin(async (tx) => {
-    await tx`delete from ${tx(table)} where ${tx(pkField)} = ${data[pkField]}`
-    await tx`insert into ${pgsql.unsafe(table)} ${pgsql(data)}`
-  })
-}
-
-async function pgUpdate(
-  table: string,
-  pkField: string,
-  data: Record<string, any>
-) {
-  await pgsql`update ${pgsql.unsafe(table)} set ${pgsql(data)} where ${pgsql.unsafe(pkField)} = ${data[pkField]}`
 }
 
 function ifdef(obj: any, snippet: any, fallback?: any) {
